task submission

+++ restaurants/views.py
@@ -59,6 +59,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if not request.user.is_authenticated:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -71,10 +73,16 @@ def restaurant_create(request):
         "form":form,
     }
     return render(request, 'create.html', context)
+def no_access(request):
+    return render(request, 'no_access.html')
+
 
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff  or request.user == restaurant.owner):
+        return redirect('noaccess')
+
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -91,6 +99,8 @@ def item_create(request, restaurant_id):
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     form = RestaurantForm(instance=restaurant_obj)
+    if not (request.user.is_staff  or request.user == restaurant_obj.owner):
+        return redirect('noaccess')
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
         if form.is_valid():
@@ -103,6 +113,8 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect('noaccess')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
